load tests refactored, to generate unique values from more wide "long" numbers range, instead of previously used "int" numbers range

+++ modules/cassandra/src/test/java/org/apache/ignite/tests/load/Generator.java
@@ -23,5 +23,5 @@ package org.apache.ignite.tests.load;
  */
 public interface Generator {
     /** */
-    public Object generate(int i);
+    public Object generate(long i);
 }
